requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
SQLAlchemy
psycopg2-binary